Rework the triage and evaluation engine and version the criteria

+++ backend/src/validators/criteria.validator.ts
@@ -4,7 +4,8 @@ export const createCriteriaSchema = z.object({
   nombre: z.string().trim().min(1).max(255),
   descripcion: z.string().trim().min(1),
   promptContext: z.string().trim().min(1),
-  peso: z.number().int().min(0).max(100),
+  // Decimal a propósito: el enunciado declara pesos de 12.5%.
+  peso: z.number().min(0).max(100),
   activo: z.boolean().optional(),
   orden: z.number().int().min(0).optional(),
 });
@@ -21,7 +22,7 @@ export const reorderCriteriaSchema = z.object({
       z.object({
         id: z.string().uuid(),
         orden: z.number().int().min(0),
-        peso: z.number().int().min(0).max(100),
+        peso: z.number().min(0).max(100),
         activo: z.boolean(),
       }),
     )
